expanded PWM and TWI modules
This update expands the PWM and TWI modules to be able to take advantage of all that the
current microntroller has to offer.  Note that these modules are NOT yet fully tested as
this would require additional hardware to test on.  It may be worth investing in a basic breakout
board (independent of any single appliation) to be able to run a unit test after each update.

  - enable up to 9 PWM pins (NOT all have been tested)
  - enable up to 3 I2C/SMBus buses as master of that bus (data collection/reading has NOT been tested)