Latency in firmware USB module
Additional experimentation showed that it took much longer than expected (~1-to-3s sometimes) to
receive an updated joystick value from the OCU.  The same PTZ firmware modified to appear as a
Boom camera showed much faster responsiveness.  The Boom USB module was then iteratively modified
until latency was reduced.  Note that this may not have solved the core issue at hand and that the overall latency was not made completely unnoticeable to the human eye (~<20ms).

  - reduce latency in firmware USB (in RXUSBDevice_ProcessMessage()) by removing a check for
    a not-busy module that failed and modifying the crapout